requirements.txt
Django==2.0
pytz==2017.3
gunicorn==19.7.1
django-heroku==0.3.1
djangorestframework==3.10.3